// src/graphql/generated.ts
export type Maybe<T> = T | null;
// ...
export type CodeOfConduct = Node & {
  __typename?: "CodeOfConduct";
  /** The body of the Code of Conduct */
  body: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  /** The key for the Code of Conduct */
  key: Scalars["String"];
  /** The formal name of the Code of Conduct */
  name: Scalars["String"];
  /** The HTTP path for this Code of Conduct */
  resourcePath: Maybe<Scalars["URI"]>;
  /** The HTTP URL for this Code of Conduct */
  url: Maybe<Scalars["URI"]>;
};
// ...
export type Enterprise = Node & {
  __typename?: "Enterprise";
  /** A URL pointing to the enterprise's public avatar. */
  avatarUrl: Scalars["URI"];
  /** Enterprise billing information visible to enterprise billing managers. */
  billingInfo: Maybe<EnterpriseBillingInfo>;
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars["DateTime"];
  /** Identifies the primary key from the database. */
  databaseId: Maybe<Scalars["Int"]>;
  /** The description of the enterprise. */
  description: Maybe<Scalars["String"]>;
  /** The description of the enterprise as HTML. */
  descriptionHTML: Scalars["HTML"];
  id: Scalars["ID"];
  /** The location of the enterprise. */
  location: Maybe<Scalars["String"]>;
  /** A list of users who are members of this enterprise. */
  members: EnterpriseMemberConnection;
  /** The name of the enterprise. */
  name: Scalars["String"];
  /** A list of organizations that belong to this enterprise. */
  organizations: OrganizationConnection;
  /** Enterprise information only visible to enterprise owners. */
  ownerInfo: Maybe<EnterpriseOwnerInfo>;
  /** The HTTP path for this enterprise. */
  resourcePath: Scalars["URI"];
  /** The URL-friendly identifier for the enterprise. */
  slug: Scalars["String"];
  /** The HTTP URL for this enterprise. */
  url: Scalars["URI"];
  /** A list of user accounts on this enterprise. */
  userAccounts: EnterpriseUserAccountConnection;
  /** Is the current viewer an admin of this enterprise? */
  viewerIsAdmin: Scalars["Boolean"];
  /** The URL of the enterprise website. */
  websiteUrl: Maybe<Scalars["URI"]>;
};
// ...
export type EnterpriseAdministratorInvitation = Node & {
  __typename?: "EnterpriseAdministratorInvitation";
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars["DateTime"];
  /** The email of the person who was invited to the enterprise. */
  email: Maybe<Scalars["String"]>;
  /** The enterprise the invitation is for. */
  enterprise: Enterprise;
  id: Scalars["ID"];
  /** The user who was invited to the enterprise. */
  invitee: Maybe<User>;
  /** The user who created the invitation. */
  inviter: Maybe<User>;
  /** The invitee's pending role in the enterprise (owner or billing_manager). */
  role: EnterpriseAdministratorRole;
};
// ...
export type GitHubMetadata = {
  __typename?: "GitHubMetadata";
  /** Returns a String that's a SHA of `github-services` */
  gitHubServicesSha: Scalars["GitObjectID"];
  /** IP addresses that users connect to for git operations */
  gitIpAddresses: Maybe<Array<Scalars["String"]>>;
  /** IP addresses that service hooks are sent from */
  hookIpAddresses: Maybe<Array<Scalars["String"]>>;
  /** IP addresses that the importer connects from */
  importerIpAddresses: Maybe<Array<Scalars["String"]>>;
  /** Whether or not users are verified */
  isPasswordAuthenticationVerifiable: Scalars["Boolean"];
  /** IP addresses for GitHub Pages' A records */
  pagesIpAddresses: Maybe<Array<Scalars["String"]>>;
};
// ...
export type License = Node & {
  __typename?: "License";
  /** The full text of the license */
  body: Scalars["String"];
  /** The conditions set by the license */
  conditions: Array<Maybe<LicenseRule>>;
  /** A human-readable description of the license */
  description: Maybe<Scalars["String"]>;
  /** Whether the license should be featured */
  featured: Scalars["Boolean"];
  /** Whether the license should be displayed in license pickers */
  hidden: Scalars["Boolean"];
  id: Scalars["ID"];
  /** Instructions on how to implement the license */
  implementation: Maybe<Scalars["String"]>;
  /** The lowercased SPDX ID of the license */
  key: Scalars["String"];
  /** The limitations set by the license */
  limitations: Array<Maybe<LicenseRule>>;
  /** The license full name specified by <https://spdx.org/licenses> */
  name: Scalars["String"];
  /** Customary short name if applicable (e.g, GPLv3) */
  nickname: Maybe<Scalars["String"]>;
  /** The permissions set by the license */
  permissions: Array<Maybe<LicenseRule>>;
  /** Whether the license is a pseudo-license placeholder (e.g., other, no-license) */
  pseudoLicense: Scalars["Boolean"];
  /** Short identifier specified by <https://spdx.org/licenses> */
  spdxId: Maybe<Scalars["String"]>;
  /** URL to the license on <https://choosealicense.com> */
  url: Maybe<Scalars["URI"]>;
};
// ...
export type MarketplaceCategory = Node & {
  __typename?: "MarketplaceCategory";
  /** The category's description. */
  description: Maybe<Scalars["String"]>;
  /** The technical description of how apps listed in this category work with GitHub. */
  howItWorks: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  /** The category's name. */
  name: Scalars["String"];
  /** How many Marketplace listings have this as their primary category. */
  primaryListingCount: Scalars["Int"];
  /** The HTTP path for this Marketplace category. */
  resourcePath: Scalars["URI"];
  /** How many Marketplace listings have this as their secondary category. */
  secondaryListingCount: Scalars["Int"];
  /** The short name of the category used in its URL. */
  slug: Scalars["String"];
  /** The HTTP URL for this Marketplace category. */
  url: Scalars["URI"];
};
// ...
export type MarketplaceListingConnection = {
  __typename?: "MarketplaceListingConnection";
  /** A list of edges. */
  edges: Maybe<Array<Maybe<MarketplaceListingEdge>>>;
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<MarketplaceListing>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars["ID"];
};
// ...
export type PageInfo = {
  __typename?: "PageInfo";
  /** When paginating forwards, the cursor to continue. */
  endCursor: Maybe<Scalars["String"]>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars["Boolean"];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars["Boolean"];
  /** When paginating backwards, the cursor to continue. */
  startCursor: Maybe<Scalars["String"]>;
};
// ...
export type Query = {
  __typename?: "Query";
  /** Look up a code of conduct by its key */
  codeOfConduct: Maybe<CodeOfConduct>;
  /** Look up a code of conduct by its key */
  codesOfConduct: Maybe<Array<Maybe<CodeOfConduct>>>;
  /** Look up an enterprise by URL slug. */
  enterprise: Maybe<Enterprise>;
  /** Look up a pending enterprise administrator invitation by invitee, enterprise and role. */
  enterpriseAdministratorInvitation: Maybe<EnterpriseAdministratorInvitation>;
  /** Look up a pending enterprise administrator invitation by invitation token. */
  enterpriseAdministratorInvitationByToken: Maybe<
    EnterpriseAdministratorInvitation
  >;
  /** Look up an open source license by its key */
  license: Maybe<License>;
  /** Return a list of known open source licenses */
  licenses: Array<Maybe<License>>;
  /** Get alphabetically sorted list of Marketplace categories */
  marketplaceCategories: Array<MarketplaceCategory>;
  /** Look up a Marketplace category by its slug. */
  marketplaceCategory: Maybe<MarketplaceCategory>;
  /** Look up a single Marketplace listing */
  marketplaceListing: Maybe<MarketplaceListing>;
  /** Look up Marketplace listings */
  marketplaceListings: MarketplaceListingConnection;
  /** Return information about the GitHub instance */
  meta: GitHubMetadata;
  /** Fetches an object given its ID. */
  node: Maybe<Node>;
  /** Lookup nodes by a list of IDs. */
  nodes: Array<Maybe<Node>>;
  /** Lookup a organization by login. */
  organization: Maybe<Organization>;
  /** The client's rate limit information. */
  rateLimit: Maybe<RateLimit>;
  /** Hack to workaround https://github.com/facebook/relay/issues/112 re-exposing the root query object */
  relay: Query;
  /** Lookup a given repository by the owner and repository name. */
  repository: Maybe<Repository>;
  /** Lookup a repository owner (ie. either a User or an Organization) by login. */
  repositoryOwner: Maybe<RepositoryOwner>;
  /** Lookup resource by a URL. */
  resource: Maybe<UniformResourceLocatable>;
  /** Perform a search across resources. */
  search: SearchResultItemConnection;
  /** GitHub Security Advisories */
  securityAdvisories: SecurityAdvisoryConnection;
  /** Fetch a Security Advisory by its GHSA ID */
  securityAdvisory: Maybe<SecurityAdvisory>;
  /** Software Vulnerabilities documented by GitHub Security Advisories */
  securityVulnerabilities: SecurityVulnerabilityConnection;
  /**
   * Look up a single Sponsors Listing
   * @deprecated `Query.sponsorsListing` will be removed. Use `Sponsorable.sponsorsListing` instead. Removal on 2020-04-01 UTC.
   */
  sponsorsListing: Maybe<SponsorsListing>;
  /** Look up a topic by name. */
  topic: Maybe<Topic>;
  /** Lookup a user by login. */
  user: Maybe<User>;
  /** The currently authenticated user. */
  viewer: User;
};
// ...
export type RateLimit = {
  __typename?: "RateLimit";
  /** The point cost for the current query counting against the rate limit. */
  cost: Scalars["Int"];
  /** The maximum number of points the client is permitted to consume in a 60 minute window. */
  limit: Scalars["Int"];
  /** The maximum number of nodes this query may return */
  nodeCount: Scalars["Int"];
  /** The number of points remaining in the current rate limit window. */
  remaining: Scalars["Int"];
  /** The time at which the current rate limit window resets in UTC epoch seconds. */
  resetAt: Scalars["DateTime"];
  /** The number of points used in the current rate limit window. */
  used: Scalars["Int"];
};
// ...
export type SearchResultItemConnection = {
  __typename?: "SearchResultItemConnection";
  /** The number of pieces of code that matched the search query. */
  codeCount: Scalars["Int"];
  /** A list of edges. */
  edges: Maybe<Array<Maybe<SearchResultItemEdge>>>;
  /** The number of issues that matched the search query. */
  issueCount: Scalars["Int"];
  /** A list of nodes. */
  nodes: Maybe<Array<Maybe<SearchResultItem>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The number of repositories that matched the search query. */
  repositoryCount: Scalars["Int"];
  /** The number of users that matched the search query. */
  userCount: Scalars["Int"];
  /** The number of wiki pages that matched the search query. */
  wikiCount: Scalars["Int"];
};

/** An edge in a connection. */
export type SearchResultItemEdge = {
  __typename?: "SearchResultItemEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  /** The item at the end of the edge. */
  node: Maybe<SearchResultItem>;
  /** Text matches on the result found. */
  textMatches: Maybe<Array<Maybe<TextMatch>>>;
};
// ...
export type RepositoryStatsFragment = { __typename?: "Repository" } & Pick<
  Repository,
  "id" | "name" | "url" | "forkCount" | "stargazerCount"
>;

export type SearchRepositoriesQueryVariables = Exact<{
  query: Scalars["String"];
  first: Scalars["Int"];
  after: Maybe<Scalars["String"]>;
}>;

export type SearchRepositoriesQuery = { __typename?: "Query" } & {
  search: { __typename?: "SearchResultItemConnection" } & {
    pageInfo: { __typename?: "PageInfo" } & Pick<
      PageInfo,
      "startCursor" | "endCursor" | "hasNextPage" | "hasPreviousPage"
    >;
    edges: Maybe<
      Array<
        { __typename?: "SearchResultItemEdge" } & Pick<
          SearchResultItemEdge,
          "cursor"
        > & {
            node: { __typename?: "Repository" } & RepositoryStatsFragment;
          }
      >
    >;
  };
};